Made compile_options CMAKE_BUILD_TYPE regex case-insensitive
In previous commit, forgot to mention that one compile option
became contingent on CMAKE_BUILD_TYPE being Release.
The regex-bases check for this is now case-insensitive to handle
the unlikely case where CMAKE_BUILD_TYPE was explicitly specified
to be Release (or relEAse, or RELeaSe, etc etc) by the user upon
running cmake.